Relocate read_choice to prompt.py
